+++ apps/simulation-study-time-series/Main.hs
@@ -31,8 +31,11 @@ import Epidemic.Types.Parameter
 import Epidemic.Types.Population (Person(..))
 import qualified Epidemic.Utility as SimUtil
 import GHC.Generics
+import Numeric.GSL.Minimization (MinimizeMethod(NMSimplex2), minimizeV)
+import Numeric.LinearAlgebra.Data (linspace, toList)
+import Numeric.LinearAlgebra.HMatrix
+
 import System.Environment (getArgs)
--- import Data.Map.Strict (Map(),fromList,toList)
 
 -- | These objects define the specifics for an inference evaluation which can be
 -- run at differing points in the simulation to understand how differing amounts
@@ -43,7 +46,7 @@ data InferenceConfiguration =
     , reconstructedTreeOutputFiles :: (FilePath, FilePath)
     , observationsOutputCsv :: FilePath
     , llhdOutputCsv :: FilePath
-    , negBinomCsv :: FilePath
+    , pointEstimatesCsv :: FilePath
     }
   deriving (Show, Generic)
 
@@ -56,7 +59,6 @@ data Configuration =
     , simulationDuration :: Time
     , simulationSizeBounds :: (Int,Int)
     , inferenceConfigurations :: [InferenceConfiguration]
-    , evaluationParameters :: [Parameters]
     , partialEvaluationOutputCsv :: FilePath
     }
   deriving (Show, Generic)
@@ -67,11 +69,18 @@ instance Json.FromJSON InferenceConfiguration
 
 type Simulation x = ReaderT Configuration (ExceptT String IO) x
 
+-- | This type is used to indicate if parameters are the true ones used in the
+-- simulation or estimates parameters.
+data ParameterKind
+  = SimulationParameters
+  | EstimatedParameters
+  deriving (Show, Eq)
+
 -- A BDSCOD simulation configuration based on the parameters in the environment.
 bdscodConfiguration = do
   simParams <- asks simulationParameters
   simDur <- asks simulationDuration
-  let bdscodConfig = SimBDSCOD.configuration simDur simParams
+  let bdscodConfig = SimBDSCOD.configuration simDur (unpackParameters simParams)
   case bdscodConfig of
     Nothing -> throwError "Could not construct BDSCOD configuration"
     (Just config) -> return config
@@ -108,19 +117,92 @@ simulatedObservations infConfig@InferenceConfiguration{..} simEvents = do
     Nothing -> throwError "Failed to simulate observations."
 
 
-
--- Run the evaluation of the log-likelihood profiles on a given set of
--- observations and write the result to file.
-evaluateLLHD :: InferenceConfiguration -> [Observation] -> Simulation ()
-evaluateLLHD InferenceConfiguration{..} obs = do
-  simParams <- asks simulationParameters
-  evalParams <- asks evaluationParameters
+-- | Evaluate the NB posterior approximation of the prevalence for a single
+-- point in parameter space and the LLHD over a list of points and write all of
+-- the results to CSV.
+generateLlhdProfileCurves :: InferenceConfiguration
+                          -> [Observation]
+                          -> (Parameters,ParameterKind,[Parameters])
+                          -> Simulation ()
+generateLlhdProfileCurves InferenceConfiguration{..} obs (singleParams,paramKind,evalParams) =
   let comma = BBuilder.charUtf8 ','
+      parametersUsed = show paramKind
+      parametersUsed' = BBuilder.stringUtf8 parametersUsed
       llhdVals = [fst $ llhdAndNB obs p initLlhdState | p <- evalParams]
-      nBVal = pure . snd $ llhdAndNB obs simParams initLlhdState
-      doublesAsString = BBuilder.toLazyByteString . mconcat . intersperse comma . map BBuilder.doubleDec
-  liftIO $ L.writeFile llhdOutputCsv (doublesAsString llhdVals)
-  liftIO $ L.writeFile negBinomCsv (Csv.encode nBVal)
+      nBVal = pure (parametersUsed,snd $ llhdAndNB obs singleParams initLlhdState)
+      doublesAsString = BBuilder.toLazyByteString . mconcat . intersperse comma . (parametersUsed':) . map BBuilder.doubleDec
+  in do
+    liftIO $ L.appendFile llhdOutputCsv (doublesAsString llhdVals)
+    liftIO $ L.appendFile pointEstimatesCsv (Csv.encode nBVal)
+
+-- | Run the evaluation of the log-likelihood profiles on a given set of
+-- observations at the parameters used to simulate the data set and write the
+-- result to file.
+evaluateLLHD :: InferenceConfiguration -> [Observation] -> Simulation ()
+evaluateLLHD infConfig obs = do
+  simParams <- asks simulationParameters -- get the actual parameters used to simulate the observations
+  evalParams <- adjustedEvaluationParameters simParams
+  generateLlhdProfileCurves infConfig obs (simParams,SimulationParameters,evalParams)
+
+-- | Estimate the parameters of the of the model and then evaluate the LLHD
+-- profiles and prevalence and append this to the file. The first value of the
+-- CSV output now describes which parameters where used to to evaluate these
+-- things.
+estimateLLHD :: InferenceConfiguration -> [Observation] -> Simulation ()
+estimateLLHD infConfig obs = do
+  simParams@(Parameters (_,deathRate,_,_,_,_)) <- asks simulationParameters
+  let schedTimes = scheduledTimes simParams
+      mleParams = estimateParameters deathRate schedTimes obs -- get the MLE estimate of the parameters
+  evalParams <- adjustedEvaluationParameters mleParams -- generate a list of evaluation parameters
+  generateLlhdProfileCurves infConfig obs (mleParams,EstimatedParameters,evalParams)
+
+-- | Use GSL to estimate the MLE based on the observations given. This uses a
+-- simplex method because it seems to be faster and more accurate than the
+-- simulated annealing.
+--
+-- __NOTE__ we fix the death rate to the true value because
+-- this is assumed to be known a priori.
+estimateParameters :: Rate -> ([Time],[Time]) -> [Observation] -> Parameters
+estimateParameters deathRate sched obs =
+  let maxIters = 500
+      desiredPrec = 1e-3
+      initBox = fromList [2,2,2,2,2]
+      energyFunc x = negate . fst $ llhdAndNB obs (vectorAsParameters deathRate sched x) initLlhdState
+      randInit = fromList [0,0,0,0,0]
+      (est,_) = minimizeV NMSimplex2 desiredPrec maxIters initBox energyFunc randInit
+  in vectorAsParameters deathRate sched est
+
+-- | Helper function for @estimateParameters@
+vectorAsParameters :: Rate -> ([Time],[Time]) -> Vector Double -> Parameters
+vectorAsParameters deathRate (rhoTimes, nuTimes) paramVec =
+  let [lnR1, lnR2, lnP1, lnR3, lnP2] = toList paramVec
+      p1 = invLogit lnP1
+      p2 = invLogit lnP2
+  in packParameters (exp lnR1, deathRate, exp lnR2, [(t,p1)|t<-rhoTimes], exp lnR3, [(t,p2)|t<-nuTimes])
+
+
+invLogit :: Double -> Probability
+invLogit a = 1 / (1 + exp (- a))
+
+logit :: Probability -> Double
+logit p = log (p / (1 - p))
+
+-- | Recenter the evaluation parametes about the parameters given.
+adjustedEvaluationParameters :: Parameters -> Simulation [Parameters]
+adjustedEvaluationParameters ps =
+  let meshSize = 100
+      lambdaMesh = toList $ linspace meshSize (1,2.5)
+      muMesh = toList $ linspace meshSize (0.05,1.5)
+      psiMesh = toList $ linspace meshSize (0.05,1.5)
+      probMesh = toList $ linspace meshSize (0.05,0.6) :: [Probability]
+      (rhoTimes,nuTimes) = scheduledTimes ps
+      rhoMesh = [Timed [(t,r) | t <- rhoTimes] | r <- probMesh]
+      omegaMesh = toList $ linspace meshSize (0.05,1.5)
+      nuMesh = [Timed [(t,n) | t <- nuTimes] | n <- probMesh]
+      apply f = map (f ps)
+      [lPs,mPs,pPs,oPs] = zipWith apply [putLambda,putMu,putPsi,putOmega] [lambdaMesh,muMesh,psiMesh,omegaMesh]
+      [rPs,nPs] = zipWith apply [putRhos,putNus] [rhoMesh,nuMesh]
+  in return $ concat [lPs,mPs,pPs,rPs,oPs,nPs]
 
 -- | Record the partial results of the LLHD and NB to a CSV at the parameters
 -- used in the simulation.
@@ -140,7 +222,9 @@ simulationStudy = do
   infConfigs <- asks inferenceConfigurations
   pObs <- zipWithM simulatedObservations infConfigs pEpi
   mapM_ (uncurry evaluateLLHD) pObs
-  partialEvaluations (snd $ head pObs)
+  mapM_ (uncurry estimateLLHD) pObs
+  let completeObs = snd $ head pObs
+  partialEvaluations completeObs
   return ()
 
 main :: IO ()