+++ examples/simulation-study-time-series/src/make-json-config.R
@@ -6,7 +6,7 @@ output_file <- "ts-config.json"
 
 simulation_duration <- 17
 
-inference_times <- seq(from = 7, to = 17, by = 2) # times to evaluate the llhd and prevalence distribution
+inference_times <- seq(from = 7, to = 17, by = 4) # times to evaluate the llhd and prevalence distribution
 
 
 birth_rate <- 1.5
@@ -33,19 +33,12 @@ inference_configuration <- function(inf_time) {
                                          inf_time),
          llhdOutputCsv = sprintf("out/llhd-evaluations-%.2f.csv",
                                  inf_time),
-         negBinomCsv = sprintf("out/final-negative-binomial-%.2f.csv",
-                               inf_time))
+         pointEstimatesCsv = sprintf("out/final-negative-binomial-%.2f.csv",
+                                     inf_time))
 }
 
 
 sim_params <- list(birth_rate, death_rate, sampling_rate, catastrophe_params, occurrence_rate, disaster_params)
-
-num_eval_steps <- 10 # the number of mesh points for plotting
-eval_lambda <- seq(from = birth_rate - 1, to = birth_rate + 1, length = num_eval_steps)
-eval_params_lambda <- map(.x = eval_lambda, .f = ~ list(.x, death_rate, sampling_rate, catastrophe_params, occurrence_rate, disaster_params))
-eval_mu <- seq(from = min(0.05, death_rate - 0.5), to = death_rate + 0.5, length = num_eval_steps)
-eval_params_mu <- map(.x = eval_mu, .f = ~ list(birth_rate, .x, sampling_rate, catastrophe_params, occurrence_rate, disaster_params))
-eval_params <- c(eval_params_lambda, eval_params_mu)
 
 
 result <- list(
@@ -54,17 +47,10 @@ result <- list(
     simulationDuration = simulation_duration + 1e-5,
     simulationSizeBounds = c(100,100000),
     inferenceConfigurations = map(inference_times, inference_configuration),
-    evaluationParameters = eval_params,
     partialEvaluationOutputCsv = "out/partial-evaluations.csv"
 )
 
 write_json(result, output_file, pretty = FALSE, auto_unbox = TRUE, digits = 7)
-
-## save a copy of the "evaluation paramters" to make it easier to plot the
-## profiles later.
-eval_df <- data.frame(parameter = rep(c("lambda", "mu"), each = num_eval_steps),
-                      value = c(eval_lambda, eval_mu))
-write.table(x = eval_df, file = "out/evaluation-parameters.csv", row.names = FALSE)
 
 ## save a copy of the true parameters so they can be read out later rather than
 ## hardcoded.